[libcxx] [test] Add casts to avoid signed/unsigned mismatch warnings on MSVC++
A bug was filed that these warnings should not be emitted as DevCom-883961. ( https://developercommunity.visualstudio.com/content/problem/883961/c4389-signedunsigned-mismatch-should-not-be-emitte.html )

+++ libcxx/test/std/algorithms/alg.modifying.operations/alg.remove/remove_copy.pass.cpp
@@ -26,7 +26,7 @@ TEST_CONSTEXPR bool test_constexpr() {
 
     auto it = std::remove_copy(std::begin(ia), std::end(ia), std::begin(ib), 5);
 
-    return std::distance(std::begin(ib), it) == (std::size(ia) - 2)   // we removed two elements
+    return std::distance(std::begin(ib), it) == static_cast<int>(std::size(ia) - 2)   // we removed two elements
         && std::none_of(std::begin(ib), it, [](int a) {return a == 5;})
         && std::all_of (it, std::end(ib),   [](int a) {return a == 0;})
            ;
